1. 修复无本地应用时右键菜单出现“开始抖动”的bug 2. 修复应用图标抖动时点击删除后崩溃的bug 3. 修复拖动应用图标时界面刷新后崩溃的bug
git-svn-id: http://svn.i-soft.com.cn/sec/idesktop-client/idesktop-client@45 e225a175-1783-44c2-b966-f136c5b5d36a

+++ src/dashboard.cpp
@@ -339,6 +339,17 @@ void Dashboard::setBgPixmap(const QString &pixText)
 
 void Dashboard::refreshDesktop()
 {
+
+    if (vdesktop->dragEvent())
+    {
+        return;
+    }
+
+    if (vdesktop->trembleState())
+    {
+        return;
+    }
+
     //delete old page.
     vdesktop->movetoFist();
     vdesktop->deleteAllIconItem();
